now sti goes backward instead of forward lol

+++ src/opt.c
@@ -6,7 +6,7 @@
 /*   By: jye <jye@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/23 22:00:01 by jye               #+#    #+#             */
-/*   Updated: 2017/03/03 22:47:47 by jye              ###   ########.fr       */
+/*   Updated: 2017/03/04 14:18:35 by root             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -282,16 +282,16 @@ void	ldi(t_vm *vm, t_process *process)
 		{
 			r[i] = get_param(vm, pc[0], (int[3]){pc[1], REG_CODE, 0});
 			pc[1] += 1;
-			r[i] = process->r[r[i] - 1];
+			r[i] = (short)process->r[r[i] - 1];
 		}
 		else if (octal == DIR_CODE)
 		{
-			r[i] = get_param(vm, pc[0], (int[3]){pc[1], DIR_CODE, 1});
+			r[i] = (short)get_param(vm, pc[0], (int[3]){pc[1], DIR_CODE, 1});
 			pc[1] += 2;
 		}
 		else
 		{
-			r[i] = get_param(vm, pc[0], (int[3]){pc[1], IND_CODE, 0});
+			r[i] = (short)get_param(vm, pc[0], (int[3]){pc[1], IND_CODE, 0});
 			pc[1] += 2;
 		}
 		offset -= 2;
@@ -316,6 +316,11 @@ void	sti(t_vm *vm, t_process *process)
 	i = 1;
 	offset = 4;
 	r[0] = get_param(vm, pc[0], (int[3]){pc[0] + 2, REG_CODE, 0});
+	/* int debug = 0; */
+	/* if (process->pc == 151) */
+	/* { */
+	/* 	debug = 1; */
+	/* } */
 	r[0] = process->r[r[0] - 1];
 	while (i < 3)
 	{
@@ -324,22 +329,34 @@ void	sti(t_vm *vm, t_process *process)
 		{
 			r[i] = get_param(vm, pc[0], (int[3]){pc[1], REG_CODE, 0});
 			pc[1] += 1;
-			r[i] = process->r[r[i] - 1];
+			r[i] = (short)process->r[r[i] - 1];
 		}
 		else if (octal == DIR_CODE)
 		{
-			r[i] = get_param(vm, pc[0], (int[3]){pc[1], DIR_CODE, 1});
+			r[i] = (short)get_param(vm, pc[0], (int[3]){pc[1], DIR_CODE, 1});
 			pc[1] += 2;
 		}
 		else
 		{
-			r[i] = get_param(vm, pc[0], (int[3]){pc[1], IND_CODE, 0});
+			r[i] = (short)get_param(vm, pc[0], (int[3]){pc[1], IND_CODE, 0});
 			pc[1] += 2;
 		}
 		offset -= 2;
 		++i;
 	}
 	st_param(vm, pc[0] + ((r[1] + r[2]) % IDX_MOD), r[0]);
+	/* if (debug) */
+	/* { */
+	/* 	printf("process->pc %u\n", process->pc); */
+	/* 	printf("vm->nb_process %u\n", vm->nb_process); */
+	/* 	printf("vm->cycle %lu\n", vm->cycle); */
+	/* 	printf("r[0]%u r[1]%d r[2]%d\n", r[0], r[1], r[2]); */
+	/* 	printf("live check %x\n", get_param(vm, pc[0], (int[3]){pc[0] - 4, DIR_CODE, 0})); */
+	/* 	printf("pc[0] + ((r[1] + r[2]) % IDX_MOD) %u\n", pc[0] + ((r[1] + r[2]) % IDX_MOD)); */
+	/* 	print_map(vm->map); */
+	/* 	fflush(stdout); */
+	/* 	abort(); */
+	/* } */
 	process->pc = pc[1];
 }
 
@@ -397,7 +414,7 @@ void	lldi(t_vm *vm, t_process *process)
 	unsigned int	pc[2];
 	unsigned int	offset;
 	int				i;
-	unsigned int	r[3];
+	int				r[3];
 	unsigned char	octal;
 
 	pc[0] = process->pc;
@@ -411,16 +428,16 @@ void	lldi(t_vm *vm, t_process *process)
 		{
 			r[i] = get_lparam(vm, pc[0], (int[3]){pc[1], REG_CODE, 0});
 			pc[1] += 1;
-			r[i] = process->r[r[i] - 1];
+			r[i] = (short)process->r[r[i] - 1];
 		}
 		else if (octal == DIR_CODE)
 		{
-			r[i] = get_lparam(vm, pc[0], (int[3]){pc[1], DIR_CODE, 1});
+			r[i] = (short)get_lparam(vm, pc[0], (int[3]){pc[1], DIR_CODE, 1});
 			pc[1] += 2;
 		}
 		else
 		{
-			r[i] = get_lparam(vm, pc[0], (int[3]){pc[1], IND_CODE, 0});
+			r[i] = (short)get_lparam(vm, pc[0], (int[3]){pc[1], IND_CODE, 0});
 			pc[1] += 2;
 		}
 		offset -= 2;
